Cleanup + better docs.
