+++ polkadot/primitives/src/v5/mod.rs
@@ -830,60 +830,6 @@ pub struct ParathreadEntry {
 	pub retries: u32,
 }
 
-/// An assignment for a parachain scheduled to be backed and included in a relay chain block.
-#[derive(Clone, Encode, Decode, PartialEq, TypeInfo, RuntimeDebug)]
-pub struct Assignment {
-	/// Assignment's ParaId
-	pub para_id: Id,
-}
-
-impl Assignment {
-	/// Create a new `Assignment`.
-	pub fn new(para_id: Id) -> Self {
-		Self { para_id }
-	}
-}
-
-/// An entry tracking a paras
-#[derive(Clone, Encode, Decode, TypeInfo, PartialEq, RuntimeDebug)]
-pub struct ParasEntry<N = BlockNumber> {
-	/// The `Assignment`
-	pub assignment: Assignment,
-	/// The number of times the entry has timed out in availability.
-	pub availability_timeouts: u32,
-	/// The block height where this entry becomes invalid.
-	pub ttl: N,
-}
-
-impl<N> ParasEntry<N> {
-	/// Return `Id` from the underlying `Assignment`.
-	pub fn para_id(&self) -> Id {
-		self.assignment.para_id
-	}
-
-	/// Create a new `ParasEntry`.
-	pub fn new(assignment: Assignment, now: N) -> Self {
-		ParasEntry { assignment, availability_timeouts: 0, ttl: now }
-	}
-}
-
-/// What is occupying a specific availability core.
-#[derive(Clone, Encode, Decode, TypeInfo, RuntimeDebug)]
-#[cfg_attr(feature = "std", derive(PartialEq))]
-pub enum CoreOccupied<N> {
-	/// The core is not occupied.
-	Free,
-	/// A paras.
-	Paras(ParasEntry<N>),
-}
-
-impl<N> CoreOccupied<N> {
-	/// Is core free?
-	pub fn is_free(&self) -> bool {
-		matches!(self, Self::Free)
-	}
-}
-
 /// A helper data-type for tracking validator-group rotations.
 #[derive(Clone, Encode, Decode, TypeInfo, RuntimeDebug)]
 #[cfg_attr(feature = "std", derive(PartialEq))]
