+++ polkadot/runtime/parachains/Cargo.toml
@@ -49,6 +49,7 @@ rand_chacha = { version = "0.3.1", default-features = false }
 static_assertions = { version = "1.1.0", optional = true }
 polkadot-parachain-primitives = { path = "../../parachain", default-features = false }
 polkadot-runtime-metrics = { path = "../metrics", default-features = false}
+polkadot-core-primitives = { path = "../../core-primitives", default-features = false }
 
 [dev-dependencies]
 futures = "0.3.21"
