+++ polkadot/runtime/parachains/src/assigner_on_demand/tests.rs
@@ -24,13 +24,11 @@ use crate::{
 		System, Test,
 	},
 	paras::{ParaGenesisArgs, ParaKind},
+	scheduler::common::Assignment,
 };
 use frame_support::{assert_noop, assert_ok, error::BadOrigin};
 use pallet_balances::Error as BalancesError;
-use primitives::{
-	v5::{Assignment, ValidationCode},
-	BlockNumber, SessionIndex,
-};
+use primitives::{v5::ValidationCode, BlockNumber, SessionIndex};
 use sp_std::collections::btree_map::BTreeMap;
 
 fn schedule_blank_para(id: ParaId, parakind: ParaKind) {
