+++ polkadot/runtime/parachains/src/scheduler/tests.rs
@@ -18,7 +18,7 @@ use super::*;
 
 use frame_support::assert_ok;
 use keyring::Sr25519Keyring;
-use primitives::{v5::Assignment, BlockNumber, SessionIndex, ValidationCode, ValidatorId};
+use primitives::{BlockNumber, SessionIndex, ValidationCode, ValidatorId};
 use sp_std::collections::{btree_map::BTreeMap, btree_set::BTreeSet};
 
 use crate::{
